refactor: move metrics out of authenticate
